refactor(bridge): rehome PR persistence into api/database structure
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
