+++ bridge/app/lib/src/bridge/api/database/tables/pull_requests_table.freezed.dart
